API change

+++ contracts/BaseBaseLock.sol
@@ -433,7 +433,7 @@ abstract contract BaseBaseLock is ERC1155WithTotals , IERC1155TokenReceiver {
     /// Start with 1, not 0, to avoid glitch with `conditionalTokens`.
     ///
     /// TODO: Use uint64 variables instead?
-    function _createCondition(address customer) internal returns (uint64) {
+    function _createCondition(address customer) internal returns (uint256) {
         uint64 _conditionId = ++maxConditionId;
         customers[_conditionId] = customer; // TODO: Be able to mint for somebody other?
         // TODO
@@ -464,7 +464,7 @@ abstract contract BaseBaseLock is ERC1155WithTotals , IERC1155TokenReceiver {
     ///
     /// FIXME: Allow to recreate only the last token in the list.
     function _recreateCondition(uint256 _condition) internal myConditional(_condition) returns (uint256) {
-        uint64 _newCondition = _createCondition(msg.sender);
+        uint256 _newCondition = _createCondition(msg.sender);
         // TODO: Store the linked list of conditional tokens for a condition.
         // TODO: misc
         // TODO: Event that related old and new condition for traders. Also relate them on-chain? (liked list? map to the first condition in the list?)
